+++ bootique-jetty/src/main/java/io/bootique/jetty/JettyModule.java
@@ -1,3 +1,22 @@
+/**
+ *  Licensed to ObjectStyle LLC under one
+ *  or more contributor license agreements.  See the NOTICE file
+ *  distributed with this work for additional information
+ *  regarding copyright ownership.  The ObjectStyle LLC licenses
+ *  this file to you under the Apache License, Version 2.0 (the
+ *  “License”); you may not use this file except in compliance
+ *  with the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  “AS IS” BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+
 package io.bootique.jetty;
 
 import com.google.inject.Binder;
@@ -10,6 +29,7 @@ import io.bootique.jetty.command.ServerCommand;
 import io.bootique.jetty.server.MappedFilterFactory;
 import io.bootique.jetty.server.MappedServletFactory;
 import io.bootique.jetty.server.ServerFactory;
+import io.bootique.jetty.server.ServletContextHandlerExtender;
 import io.bootique.jetty.servlet.DefaultServletEnvironment;
 import io.bootique.jetty.servlet.ServletEnvironment;
 import io.bootique.log.BootLogger;
@@ -84,13 +104,15 @@ public class JettyModule extends ConfigModule {
                         Set<MappedFilter> mappedFilters,
                         Set<EventListener> listeners,
                         Set<MappedListener> mappedListeners,
+                        Set<ServletContextHandlerExtender> contextHandlerExtenders,
                         BootLogger bootLogger,
                         ShutdownManager shutdownManager) {
 
         Server server = factory.createServer(
                 allServlets(servlets, mappedServlets),
                 allFilters(filters, mappedFilters),
-                allListeners(listeners, mappedListeners));
+                allListeners(listeners, mappedListeners),
+                contextHandlerExtenders);
 
         shutdownManager.addShutdownHook(() -> {
             bootLogger.trace(() -> "stopping Jetty...");