+++ bootique-jetty/src/test/java/io/bootique/jetty/StaticResourcesIT.java
@@ -1,3 +1,22 @@
+/**
+ *  Licensed to ObjectStyle LLC under one
+ *  or more contributor license agreements.  See the NOTICE file
+ *  distributed with this work for additional information
+ *  regarding copyright ownership.  The ObjectStyle LLC licenses
+ *  this file to you under the Apache License, Version 2.0 (the
+ *  “License”); you may not use this file except in compliance
+ *  with the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  “AS IS” BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+
 package io.bootique.jetty;
 
 import io.bootique.BQCoreModule;
@@ -15,6 +34,25 @@ import java.net.MalformedURLException;
 import static org.junit.Assert.assertEquals;
 
 public class StaticResourcesIT {
+
+    public static final String COPYRIGHT = "<!--\n" +
+            "   Licensed to ObjectStyle LLC under one\n" +
+            "   or more contributor license agreements.  See the NOTICE file\n" +
+            "   distributed with this work for additional information\n" +
+            "   regarding copyright ownership.  The ObjectStyle LLC licenses\n" +
+            "   this file to you under the Apache License, Version 2.0 (the\n" +
+            "   “License”); you may not use this file except in compliance\n" +
+            "   with the License.  You may obtain a copy of the License at\n" +
+            "\n" +
+            "     http://www.apache.org/licenses/LICENSE-2.0\n" +
+            "\n" +
+            "   Unless required by applicable law or agreed to in writing,\n" +
+            "   software distributed under the License is distributed on an\n" +
+            "   “AS IS” BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY\n" +
+            "   KIND, either express or implied.  See the License for the\n" +
+            "   specific language governing permissions and limitations\n" +
+            "   under the License.\n" +
+            "  -->\n";
 
     @Rule
     public BQTestFactory testFactory = new BQTestFactory().autoLoadModules();
@@ -113,7 +151,7 @@ public class StaticResourcesIT {
 
         Response r = base.path("/").request().get();
         assertEquals(Status.OK.getStatusCode(), r.getStatus());
-        assertEquals("<html><body><h2>Hi!</h2></body></html>", r.readEntity(String.class));
+        assertEquals(COPYRIGHT + "<html><body><h2>Hi!</h2></body></html>", r.readEntity(String.class));
     }
 
     @Test
@@ -131,7 +169,7 @@ public class StaticResourcesIT {
 
         Response r = base.path("/").request().get();
         assertEquals(Status.OK.getStatusCode(), r.getStatus());
-        assertEquals("<html><body><h2>Hi!</h2></body></html>", r.readEntity(String.class));
+        assertEquals(COPYRIGHT + "<html><body><h2>Hi!</h2></body></html>", r.readEntity(String.class));
     }
 
     @Test
@@ -174,7 +212,7 @@ public class StaticResourcesIT {
 
         Response r3 = base.path("/sub2/").request().get();
         assertEquals(Status.OK.getStatusCode(), r3.getStatus());
-        assertEquals("<html><body><h2>2</h2></body></html>", r3.readEntity(String.class));
+        assertEquals(COPYRIGHT + "<html><body><h2>2</h2></body></html>", r3.readEntity(String.class));
     }
 
 }